+++ src/heart/css/header.module.css
@@ -9,7 +9,8 @@ header{
     justify-content: space-between;
 }
 .title{
-    font-size:5vw;
+    text-align: center;
+    font-size:4vw;
 }
 section{
     margin-top: 2vh;
