+++ src/location/css/Location.module.css
@@ -1,7 +1,21 @@
 .container{
-    display:flex;
+    width:100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 10px;
+    background-color: white;
+    border-radius: 25px;
+    padding-bottom:10vh;
 }
-.map{
-    width:40vw;
-    height:40vh;
-} 
+.title{
+    font-size: 5vw;
+    text-align: center;
+}
+hr{
+    width:100%;
+}
+.text{
+    font-size:20px;
+    text-align:center;
+}
