+++ src/location/css/LocationMap.module.css
@@ -1,0 +1,4 @@
+.map{
+    width:100%;
+    height:300px;
+}
