+++ src/login/css/Login.module.css
@@ -9,7 +9,7 @@
     padding-bottom: 10vh;
 }
 .text{
-    font-size: 25px;
+    font-size: 15px;
     display: flex;
     flex-direction: column;
     justify-content: center;
