+++ src/search/css/Searchbar.module.css
@@ -13,6 +13,9 @@
     color:darkgray;
     text-decoration:none;
 }
+.search:focus{
+  outline:none;
+}
 .submit{
     height:100%;
     border-radius:10px;
